chore(by): move locator to 'by' directory
reference #11
